Commit refactoring and add asset
